some refactoring but still doesnt load
