+++ service_application_package/__init__.py
@@ -3,6 +3,7 @@ from flask_sqlalchemy import SQLAlchemy
 from flask_bcrypt import Bcrypt
 from flask_login import LoginManager
 from flask_mail import Mail
+from flask_socketio import SocketIO
 from service_application_package.config import Config
 
 db = SQLAlchemy()
@@ -21,6 +22,8 @@ def create_app(config_class=Config):
     bcrypt.init_app(application)
     login_manager.init_app(application)
     mail.init_app(application)
+    from service_application_package.chat.routes import socketio
+    socketio.init_app(application)
 
     from service_application_package.users.routes import users
     from service_application_package.projects.routes import projects
