+++ service_application_package/chat/__init__.py
@@ -1,3 +1,12 @@
 from flask import Flask
 import redis
+import os
+
+
+url = os.environ.get('REDISCLOUD_URL')
+port = os.environ.get('REDIS_PORT')
+password = os.environ.get('REDIS_PW')
+
+
+# myredis = redis.Redis(host=url,port=port,password=password)
 myredis = redis.Redis(host='localhost',port=6379,db=0)
